DUX-3507: support curl logs endpoint
This is required to translate between the "vendor neutral" /v1/logs/get API that is normal HTTP, invented in https://github.com/facebook/buck2/pull/770 and the underlying manifold-like storage.

We have to support this HTTP API since buck2's alternative to that is manifold get bucket/file which we don't have at runtime in buck2.