[DUX-3531] [DUX-3572] implement a cleanup task for old files (#25)
* DUX-3531: implement a cleanup task for old files

These don't have to do any additional cleanup besides deleting the
records in the DB as a trigger will delete the linked blob objects. Will
have to be revisited if/when we bother tiering stuff out to S3
(DUX-3532).

* DUX-3572: Fix the fatal error reporter

This also fixes the previous commit which didn't ensure that background
tasks get shut down when the server shuts down.